+++ .gitignore
@@ -7,5 +7,3 @@
 dist/*
 docs/source/gallery/**
 .pre-commit-config.yaml
-graphix/sim/graphix.code-workspace
-graphix/sim/graphix.code-workspace
