+++ graphix/version.py
@@ -1,2 +1,1 @@
 __version__ = "0.2.5"
-
