+++ tests/random_circuit.py
@@ -1,5 +1,4 @@
 import numpy as np
-
 from graphix.transpiler import Circuit
